+++ doc/conf.py
@@ -60,7 +60,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = u'Parallel-SSH'
-copyright = u'2014-2017, P Kittenis'
+copyright = u'2014-2020, P Kittenis'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the